requirements.txt
Django==2.1.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
nose==1.3.7
pytz==2017.3
uWSGI==2.0.15
